refactor: validate signature outside of async-lock
